requirements.txt
feedparser
requests
dropbox
flask
PyPDF2
python-dotenv
